# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AI Visibility Demo API
  version: 0.1.0
  description: Contract-first API surface for config, visibility, and insights services.
servers:
  - url: http://localhost:8001
    description: config-service
  - url: http://localhost:8002
    description: visibility-service
  - url: http://localhost:8003
    description: insights-service
paths:
...
  /api/v1/brands:
    get:
      operationId: listBrands
      responses:
        "200":
          description: Brand list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"
    post:
      operationId: createBrand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrandRequest"
      responses:
        "201":
          description: Created brand.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
  /api/v1/brands/{brand_id}/competitors:
    get:
      operationId: listCompetitors
      parameters:
        - name: brand_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Competitor list for a brand.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Competitor"
    post:
      operationId: createCompetitor
      parameters:
        - name: brand_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompetitorRequest"
      responses:
        "201":
          description: Created competitor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competitor"
  /api/v1/brands/{brand_id}/products:
    get:
      operationId: listProducts
      parameters:
        - name: brand_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Product list for a brand.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      parameters:
        - name: brand_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Created product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /api/v1/prompt-sets:
    get:
      operationId: listPromptSets
      parameters:
        - name: brand_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Prompt set list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromptSet"
    post:
      operationId: createPromptSet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromptSetRequest"
      responses:
        "201":
          description: Created prompt set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptSet"
  /api/v1/prompts:
    get:
      operationId: listPrompts
      parameters:
        - name: prompt_set_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Prompt list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prompt"
    post:
      operationId: createPrompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromptRequest"
      responses:
        "201":
          description: Created prompt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
  /api/v1/prompts/{prompt_id}/versions:
    post:
      operationId: createPromptVersion
      parameters:
        - name: prompt_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromptVersionRequest"
      responses:
        "201":
          description: Created active prompt version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
  /api/v1/providers:
    get:
      operationId: listProviders
      responses:
        "200":
          description: Provider list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
    post:
      operationId: createProvider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProviderRequest"
      responses:
        "201":
          description: Created provider.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
  /api/v1/provider-credentials:
    get:
      operationId: listProviderCredentials
      responses:
        "200":
          description: Redacted provider credential metadata list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderCredential"
    post:
      operationId: createProviderCredential
      description: Saves a provider token as a write-only secret and returns redacted metadata.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProviderCredentialRequest"
      responses:
        "201":
          description: Redacted credential metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderCredential"
  /api/v1/rate-limits:
    get:
      operationId: listRateLimitPolicies
      responses:
        "200":
          description: Rate limit policy list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RateLimitPolicy"
    post:
      operationId: createRateLimitPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRateLimitPolicyRequest"
      responses:
        "201":
          description: Created rate limit policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitPolicy"
  /api/v1/models:
    get:
      operationId: listModels
      parameters:
        - name: provider_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Model registry entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModelRegistry"
    post:
      operationId: createModelRegistryEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModelRegistryRequest"
      responses:
        "201":
          description: Created model registry entry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelRegistry"
  /api/v1/providers/{provider_id}/models/sync:
    post:
      operationId: syncProviderModels
      parameters:
        - name: provider_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Provider model registry sync result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelSyncResponse"
...
components:
  schemas:
...
    Brand:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        website_url:
          type: string
          nullable: true
    CreateBrandRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        website_url:
          type: string
    Competitor:
      type: object
      required: [id, brand_id, name]
      properties:
        id:
          type: string
          format: uuid
        brand_id:
          type: string
          format: uuid
        name:
          type: string
        website_url:
          type: string
          nullable: true
    CreateCompetitorRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        website_url:
          type: string
          nullable: true
    Product:
      type: object
      required: [id, brand_id, name, description]
      properties:
        id:
          type: string
          format: uuid
        brand_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    CreateProductRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
          default: ""
    PromptSet:
      type: object
      required: [id, brand_id, name, description, is_active]
      properties:
        id:
          type: string
          format: uuid
        brand_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        is_active:
          type: boolean
    CreatePromptSetRequest:
      type: object
      required: [brand_id, name]
      properties:
        brand_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          default: ""
        is_active:
          type: boolean
          default: true
    Prompt:
      type: object
      required: [id, prompt_set_id, name, intent, active_version]
      properties:
        id:
          type: string
          format: uuid
        prompt_set_id:
          type: string
          format: uuid
        name:
          type: string
        intent:
          type: string
        active_version:
          $ref: "#/components/schemas/PromptVersion"
    PromptVersion:
      type: object
      required: [id, version, prompt_text]
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
        prompt_text:
          type: string
    CreatePromptRequest:
      type: object
      required: [prompt_set_id, name, intent, prompt_text]
      properties:
        prompt_set_id:
          type: string
          format: uuid
        name:
          type: string
        intent:
          type: string
        prompt_text:
          type: string
    CreatePromptVersionRequest:
      type: object
      required: [prompt_text]
      properties:
        prompt_text:
          type: string
    Provider:
      type: object
      required: [id, provider_key, display_name, provider_kind, is_active]
      properties:
        id:
          type: string
          format: uuid
        provider_key:
          type: string
        display_name:
          type: string
        provider_kind:
          type: string
        is_active:
          type: boolean
    CreateProviderRequest:
      type: object
      required: [provider_key, display_name, provider_kind]
      properties:
        provider_key:
          type: string
        display_name:
          type: string
        provider_kind:
          type: string
        is_active:
          type: boolean
          default: true
    CreateProviderCredentialRequest:
      type: object
      required: [provider_id, label, token]
      properties:
        provider_id:
          type: string
          format: uuid
        label:
          type: string
        token:
          type: string
          writeOnly: true
    ProviderCredential:
      type: object
      required: [id, provider_id, label, status, redacted_fingerprint]
      properties:
        id:
          type: string
          format: uuid
        provider_id:
          type: string
          format: uuid
        label:
          type: string
        status:
          type: string
          enum: [active, disabled, invalid]
        redacted_fingerprint:
          type: string
        last_tested_at:
          type: string
          format: date-time
          nullable: true
    CreateRateLimitPolicyRequest:
      type: object
      required: [provider_id, max_concurrent_requests, requests_per_minute]
      properties:
        provider_id:
          type: string
          format: uuid
        model_id:
          type: string
          nullable: true
        max_concurrent_requests:
          type: integer
          minimum: 1
        requests_per_minute:
          type: integer
          minimum: 1
        tokens_per_minute:
          type: integer
          minimum: 1
          nullable: true
        min_delay_ms:
          type: integer
          minimum: 0
          default: 0
        max_retries:
          type: integer
          minimum: 0
          default: 3
        backoff_base_ms:
          type: integer
          minimum: 0
          default: 1000
        backoff_max_ms:
          type: integer
          minimum: 0
          default: 60000
    RateLimitPolicy:
      type: object
      required: [provider_id, max_concurrent_requests, requests_per_minute, min_delay_ms, max_retries, backoff_base_ms, backoff_max_ms]
      properties:
        id:
          type: string
          format: uuid
        provider_id:
          type: string
          format: uuid
        model_id:
          type: string
          nullable: true
        max_concurrent_requests:
          type: integer
        requests_per_minute:
          type: integer
        tokens_per_minute:
          type: integer
          nullable: true
        min_delay_ms:
          type: integer
        max_retries:
          type: integer
        backoff_base_ms:
          type: integer
        backoff_max_ms:
          type: integer
    ModelRegistry:
      type: object
      required: [id, provider_id, model_id, display_name, is_available, enabled_for_visibility, capability_json]
      properties:
        id:
          type: string
          format: uuid
        provider_id:
          type: string
          format: uuid
        model_id:
          type: string
        display_name:
          type: string
        owned_by:
          type: string
          nullable: true
        is_available:
          type: boolean
        enabled_for_visibility:
          type: boolean
        rate_limit_policy_id:
          type: string
          format: uuid
          nullable: true
        capability_json:
          type: object
          additionalProperties: true
    CreateModelRegistryRequest:
      type: object
      required: [provider_id, model_id, display_name]
      properties:
        provider_id:
          type: string
          format: uuid
        model_id:
          type: string
        display_name:
          type: string
        owned_by:
          type: string
          nullable: true
        is_available:
          type: boolean
          default: true
        enabled_for_visibility:
          type: boolean
          default: false
        rate_limit_policy_id:
          type: string
          format: uuid
          nullable: true
        capability_json:
          type: object
          additionalProperties: true
    ModelSyncResponse:
      type: object
      required:
        - provider_id
        - provider_key
        - discovered_count
        - created_count
        - updated_count
        - unavailable_count
        - models
      properties:
        provider_id:
          type: string
          format: uuid
        provider_key:
          type: string
        discovered_count:
          type: integer
        created_count:
          type: integer
        updated_count:
          type: integer
        unavailable_count:
          type: integer
        models:
          type: array
          items:
            $ref: "#/components/schemas/ModelRegistry"
...